new: add "--dcs" switch to "!mission weather"
In regular mode, "!mission weather" will print a textual description
of the weather in the running mission.

With the "--dcs" switch, ESST will print the weather as it actually is
in DCS, outputting the raw values for all managed weather parameters
(such as wind direction at 2000m in DCS format, fog visibility, etc.).